+++ evm/src/main/java/org/hyperledger/besu/evm/gascalculator/IstanbulGasCalculator.java
@@ -14,7 +14,6 @@
  */
 package org.hyperledger.besu.evm.gascalculator;
 
-import org.hyperledger.besu.evm.Gas;
 import org.hyperledger.besu.evm.account.Account;
 
 import org.apache.tuweni.bytes.Bytes;
@@ -22,24 +21,24 @@ import org.apache.tuweni.units.bigints.UInt256;
 
 public class IstanbulGasCalculator extends PetersburgGasCalculator {
 
-  private static final Gas TX_DATA_ZERO_COST = Gas.of(4L);
-  private static final Gas ISTANBUL_TX_DATA_NON_ZERO_COST = Gas.of(16L);
-  private static final Gas TX_BASE_COST = Gas.of(21_000L);
+  private static final long TX_DATA_ZERO_COST = 4L;
+  private static final long ISTANBUL_TX_DATA_NON_ZERO_COST = 16L;
+  private static final long TX_BASE_COST = 21_000L;
 
-  private static final Gas SLOAD_GAS = Gas.of(800);
-  private static final Gas BALANCE_OPERATION_GAS_COST = Gas.of(700);
-  private static final Gas EXTCODE_HASH_COST = Gas.of(700);
+  private static final long SLOAD_GAS = 800L;
+  private static final long BALANCE_OPERATION_GAS_COST = 700L;
+  private static final long EXTCODE_HASH_COST = 700L;
 
-  private static final Gas SSTORE_SET_GAS = Gas.of(20_000);
-  private static final Gas SSTORE_RESET_GAS = Gas.of(5_000);
-  private static final Gas SSTORE_CLEARS_SCHEDULE = Gas.of(15_000);
+  private static final long SSTORE_SET_GAS = 20_000L;
+  private static final long SSTORE_RESET_GAS = 5_000L;
+  private static final long SSTORE_CLEARS_SCHEDULE = 15_000L;
 
-  private static final Gas SSTORE_SET_GAS_LESS_SLOAD_GAS = SSTORE_SET_GAS.minus(SLOAD_GAS);
-  private static final Gas SSTORE_RESET_GAS_LESS_SLOAD_GAS = SSTORE_RESET_GAS.minus(SLOAD_GAS);
-  private static final Gas NEGATIVE_SSTORE_CLEARS_SCHEDULE = Gas.ZERO.minus(SSTORE_CLEARS_SCHEDULE);
+  private static final long SSTORE_SET_GAS_LESS_SLOAD_GAS = SSTORE_SET_GAS - SLOAD_GAS;
+  private static final long SSTORE_RESET_GAS_LESS_SLOAD_GAS = SSTORE_RESET_GAS - SLOAD_GAS;
+  private static final long NEGATIVE_SSTORE_CLEARS_SCHEDULE = -SSTORE_CLEARS_SCHEDULE;
 
   @Override
-  public Gas transactionIntrinsicGasCost(final Bytes payload, final boolean isContractCreation) {
+  public long transactionIntrinsicGasCost(final Bytes payload, final boolean isContractCreation) {
     int zeros = 0;
     for (int i = 0; i < payload.size(); i++) {
       if (payload.get(i) == 0) {
@@ -48,17 +47,15 @@ public class IstanbulGasCalculator extends PetersburgGasCalculator {
     }
     final int nonZeros = payload.size() - zeros;
 
-    Gas cost =
-        TX_BASE_COST
-            .plus(TX_DATA_ZERO_COST.times(zeros))
-            .plus(ISTANBUL_TX_DATA_NON_ZERO_COST.times(nonZeros));
+    final long cost =
+        TX_BASE_COST + (TX_DATA_ZERO_COST * zeros) + (ISTANBUL_TX_DATA_NON_ZERO_COST * nonZeros);
 
-    return isContractCreation ? cost.plus(txCreateExtraGasCost()) : cost;
+    return isContractCreation ? (cost + txCreateExtraGasCost()) : cost;
   }
 
   @Override
   // As per https://eips.ethereum.org/EIPS/eip-2200
-  public Gas calculateStorageCost(
+  public long calculateStorageCost(
       final Account account, final UInt256 key, final UInt256 newValue) {
 
     final UInt256 currentValue = account.getStorageValue(key);
@@ -76,24 +73,24 @@ public class IstanbulGasCalculator extends PetersburgGasCalculator {
 
   @Override
   // As per https://eips.ethereum.org/EIPS/eip-2200
-  public Gas calculateStorageRefundAmount(
+  public long calculateStorageRefundAmount(
       final Account account, final UInt256 key, final UInt256 newValue) {
 
     final UInt256 currentValue = account.getStorageValue(key);
     if (currentValue.equals(newValue)) {
-      return Gas.ZERO;
+      return 0L;
     } else {
       final UInt256 originalValue = account.getOriginalStorageValue(key);
       if (originalValue.equals(currentValue)) {
         if (originalValue.isZero()) {
-          return Gas.ZERO;
+          return 0L;
         } else if (newValue.isZero()) {
           return SSTORE_CLEARS_SCHEDULE;
         } else {
-          return Gas.ZERO;
+          return 0L;
         }
       } else {
-        Gas refund = Gas.ZERO;
+        long refund = 0L;
         if (!originalValue.isZero()) {
           if (currentValue.isZero()) {
             refund = NEGATIVE_SSTORE_CLEARS_SCHEDULE;
@@ -104,8 +101,8 @@ public class IstanbulGasCalculator extends PetersburgGasCalculator {
 
         if (originalValue.equals(newValue)) {
           refund =
-              refund.plus(
-                  originalValue.isZero()
+              refund
+                  + (originalValue.isZero()
                       ? SSTORE_SET_GAS_LESS_SLOAD_GAS
                       : SSTORE_RESET_GAS_LESS_SLOAD_GAS);
         }
@@ -116,24 +113,24 @@ public class IstanbulGasCalculator extends PetersburgGasCalculator {
 
   @Override
   // As per https://eips.ethereum.org/EIPS/eip-1884
-  public Gas getSloadOperationGasCost() {
+  public long getSloadOperationGasCost() {
     return SLOAD_GAS;
   }
 
   @Override
   // As per https://eips.ethereum.org/EIPS/eip-1884
-  public Gas getBalanceOperationGasCost() {
+  public long getBalanceOperationGasCost() {
     return BALANCE_OPERATION_GAS_COST;
   }
 
   @Override
   // As per https://eips.ethereum.org/EIPS/eip-1884
-  public Gas extCodeHashOperationGasCost() {
+  public long extCodeHashOperationGasCost() {
     return EXTCODE_HASH_COST;
   }
 
   @Override
-  public Gas getMaximumTransactionCost(final int size) {
-    return TX_BASE_COST.plus(ISTANBUL_TX_DATA_NON_ZERO_COST.times(size));
+  public long getMaximumTransactionCost(final int size) {
+    return TX_BASE_COST + (ISTANBUL_TX_DATA_NON_ZERO_COST * size);
   }
 }