+++ evm/src/test/java/org/hyperledger/besu/evm/precompile/BLS12MapFpToG1PrecompiledContractTest.java
@@ -20,7 +20,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
-import org.hyperledger.besu.evm.Gas;
 import org.hyperledger.besu.evm.frame.MessageFrame;
 
 import java.io.IOException;
@@ -38,7 +37,7 @@ import org.mockito.ArgumentCaptor;
 @RunWith(Parameterized.class)
 public class BLS12MapFpToG1PrecompiledContractTest {
 
-  BLS12MapFpToG1PrecompiledContract contract = new BLS12MapFpToG1PrecompiledContract();
+  final BLS12MapFpToG1PrecompiledContract contract = new BLS12MapFpToG1PrecompiledContract();
 
   private final MessageFrame messageFrame = mock(MessageFrame.class);
 
@@ -85,7 +84,7 @@ public class BLS12MapFpToG1PrecompiledContractTest {
       assertThat(expectedComputation.size()).isZero();
     } else {
       assertThat(actualComputation).isEqualTo(expectedComputation);
-      assertThat(contract.gasRequirement(input)).isEqualTo(Gas.of(Long.parseLong(expectedGasUsed)));
+      assertThat(contract.gasRequirement(input)).isEqualTo(Long.parseLong(expectedGasUsed));
     }
   }
 }